iadapt to abstract class changes
