+++ textgrid_convert/textgridtools.py
@@ -101,6 +101,21 @@ Object class = "TextGrid"
 0
 """
 
+def to_textgrid_time(timestamp, split_char="."):
+    """
+    Output needs to be in mili seconds, round to 2
+    Args:
+    00:52:58,579            timestamp (str)
+    Returns 
+    """
+    if not isinstance(timestamp, str):
+        log.debug("Converting timestamp {} to string".format(timestamp))
+        timestamp = str(timestamp)
+    time, ms = timestamp.split(split_char)
+    hours, mins, secs = [float(i) for i in time.split(":")]
+    fulltime = (hours * 3600000) + (mins * 60000) + (secs * 1000) + float(ms)
+    fulltime = fulltime / 1000
+    return fulltime
 
 TEXTGRID_HEADER = """
 File type = "ooTextFile"
